multimedia keys handling
new signal 'MultiMediaKey' emitted when a multimedia
(play-pause, next, previous, volume up/down) key is pressed

They are captured via a global event tap to swallow requested
multimedia keys (so that default applications don't get them).

The eventtap runs in a separate thread to not disturb the native
menus: they get laggy and can even freeze the screen if the
eventtap is added to the main run loop.